break subcommands into subdir, stub out enable and disable
