+++ src/commands/status.rs
@@ -1,0 +1,78 @@
+/*
+ * Author: Dave Eddy <dave@daveeddy.com>
+ * Date: February 15, 2022
+ * License: MIT
+ */
+
+//! `vsv status` subcommand
+
+use anyhow::{Context, Result};
+use rayon::prelude::*;
+use yansi::Style;
+
+use crate::config::Config;
+use crate::runit;
+use crate::service::Service;
+use crate::{utils, utils::verbose};
+
+/// Handle `vsv status` or `vsv` without a subcommand given.
+pub fn do_status(cfg: &Config) -> Result<()> {
+    // find all services
+    let services = runit::get_services(&cfg.svdir, cfg.log, cfg.get_filter())
+        .with_context(|| {
+        format!("failed to list services in {:?}", cfg.svdir)
+    })?;
+
+    // loop each service found (just gather data here, can be done in parallel)
+    let services: Vec<(Service, Vec<String>)> = services
+        .par_iter()
+        .map(|service| {
+            Service::from_runit_service(
+                service,
+                cfg.tree,
+                &cfg.proc_path,
+                &cfg.pstree_prog,
+            )
+        })
+        .collect();
+
+    // print gathared data
+    let bold_style = Style::default().bold();
+
+    println!();
+    verbose!(cfg, "found {} services in {:?}", services.len(), cfg.svdir);
+    println!(
+        "{}",
+        utils::format_status_line(
+            ("", &bold_style),
+            ("SERVICE", &bold_style),
+            ("STATE", &bold_style),
+            ("ENABLED", &bold_style),
+            ("PID", &bold_style),
+            ("COMMAND", &bold_style),
+            ("TIME", &bold_style),
+        )
+    );
+
+    // print each service found
+    for (service, messages) in services {
+        println!("{}", service);
+
+        // print pstree if applicable
+        if cfg.tree {
+            let tree_s = service.format_pstree();
+            println!("{}", tree_s);
+        }
+
+        // print any verbose messages/warnings generated by the service
+        for message in messages {
+            verbose!(cfg, "{}", message);
+        }
+    }
+
+    if !cfg.tree {
+        println!();
+    }
+
+    Ok(())
+}
