+++ src/main.rs
@@ -8,14 +8,11 @@
 //!
 //! Original: <https://github.com/bahamas10/vsv>
 
-use std::env;
-
-use anyhow::{anyhow, Context, Result};
-use clap::crate_name;
-use rayon::prelude::*;
-use yansi::{Color, Paint, Style};
+use anyhow::Result;
+use yansi::{Color, Paint};
 
 mod arguments;
+mod commands;
 mod config;
 mod die;
 mod runit;
@@ -25,136 +22,6 @@ mod utils;
 use arguments::Commands;
 use config::Config;
 use die::die;
-use service::Service;
-
-/// A `println!()`-like macro that will only print if `-v` is set.
-macro_rules! verbose {
-    ($cfg:expr, $fmt:expr $(, $args:expr )* $(,)? ) => {
-        if $cfg.verbose > 0 {
-            let s = format!($fmt $(, $args)*);
-            eprintln!(">  {}", Style::default().dimmed().paint(s));
-        }
-    };
-}
-
-/// Handle `vsv <any-non-matching-command>`.
-fn do_external(cfg: &Config, args: &[String]) -> Result<()> {
-    assert!(!args.is_empty());
-
-    let sv = cfg.sv_prog.to_owned();
-
-    if args.len() < 2 {
-        return Err(anyhow!("argument expected for '{} {}'", sv, args[0]));
-    }
-
-    // format arguments
-    let args_s = args.join(" ");
-
-    // set SVDIR env to match what user wanted
-    env::set_var(config::ENV_SVDIR, &cfg.svdir);
-
-    println!(
-        "[{}] {}",
-        crate_name!(),
-        Color::Cyan.paint(format!(
-            "Running {} command ({}={:?} {} {})",
-            sv,
-            config::ENV_SVDIR,
-            &cfg.svdir,
-            sv,
-            &args_s
-        ))
-    );
-
-    // run the actual program
-    let status = utils::run_program_get_status(&sv, args);
-
-    // check the process status
-    match status {
-        Ok(status) => {
-            let code = status.code().unwrap_or(-1);
-            let color = match code {
-                0 => Color::Green,
-                _ => Color::Red,
-            };
-
-            // print exit code
-            println!(
-                "[{}] {}",
-                crate_name!(),
-                color.paint(format!("[{} {}] exit code {}", sv, &args_s, code))
-            );
-
-            match code {
-                0 => Ok(()),
-                _ => Err(anyhow!("call to {} failed", sv)),
-            }
-        }
-        Err(err) => Err(anyhow!("failed to execute {}: {}", sv, err)),
-    }
-}
-
-/// Handle `vsv status` or `vsv` without a subcommand given.
-fn do_status(cfg: &Config) -> Result<()> {
-    // find all services
-    let services = runit::get_services(&cfg.svdir, cfg.log, cfg.get_filter())
-        .with_context(|| {
-        format!("failed to list services in {:?}", cfg.svdir)
-    })?;
-
-    // loop each service found (just gather data here, can be done in parallel)
-    let services: Vec<(Service, Vec<String>)> = services
-        .par_iter()
-        .map(|service| {
-            Service::from_runit_service(
-                service,
-                cfg.tree,
-                &cfg.proc_path,
-                &cfg.pstree_prog,
-            )
-        })
-        .collect();
-
-    // print gathared data
-    let bold_style = Style::default().bold();
-
-    println!();
-    verbose!(cfg, "found {} services in {:?}", services.len(), cfg.svdir);
-    println!(
-        "{}",
-        utils::format_status_line(
-            ("", &bold_style),
-            ("SERVICE", &bold_style),
-            ("STATE", &bold_style),
-            ("ENABLED", &bold_style),
-            ("PID", &bold_style),
-            ("COMMAND", &bold_style),
-            ("TIME", &bold_style),
-        )
-    );
-
-    // print each service found
-    for (service, messages) in services {
-        println!("{}", service);
-
-        // print pstree if applicable
-        if cfg.tree {
-            let tree_s = service.format_pstree();
-            println!("{}", tree_s);
-        }
-
-        // print any verbose messages/warnings generated by the service
-        for message in messages {
-            verbose!(cfg, "{}", message);
-        }
-    }
-
-    if !cfg.tree {
-        println!();
-    }
-
-    Ok(())
-}
 
 /// Main wrapped to return a result.
 fn do_main() -> Result<()> {
@@ -165,14 +32,25 @@ fn do_main() -> Result<()> {
     let args = arguments::parse();
     let cfg = Config::from_args(&args)?;
 
+    // toggle color if the user wants it or the env dictates
     if cfg.colorize {
         Paint::enable();
     }
 
     // figure out subcommand to run
     match &args.command {
-        Some(Commands::External(args)) => do_external(&cfg, args),
-        None | Some(Commands::Status { .. }) => do_status(&cfg),
+        None | Some(Commands::Status { .. }) => {
+            commands::status::do_status(&cfg)
+        }
+        Some(Commands::Enable { services }) => {
+            commands::enable_disable::do_enable(&cfg, services)
+        }
+        Some(Commands::Disable { services }) => {
+            commands::enable_disable::do_disable(&cfg, services)
+        }
+        Some(Commands::External(args)) => {
+            commands::external::do_external(&cfg, args)
+        }
     }
 }
